Backend/app/requirement.txt
fastapi==0.88.0
uvicorn ==0.20.0
tortoise-orm==0.19.2
aerich==0.7.1
async def main():
    # قائمة بجميع أوامر الترحيل
    migration_commands = [
       " aerich init -t aerich_confing.ORM_MIGRATION_SQL",
          "aerich init-db  ",
       "aerich migrate",
        " aerich upgrade",
       # "aerich migrate --name your_migration_name",
        #"aerich downgrade",
        #"aerich downgrade --name your_migration_name",
        #"aerich history",
    ]
